requirements.txt
Django==4.2.13
djangorestframework==3.15.1
psycopg2-binary==2.9.9
django-redis==5.4.0
redis==5.0.4
python-dotenv==1.0.1
celery==5.4.0
gunicorn==22.0.0
